requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
structlog>=24.0.0
anthropic>=0.25.0
PyGithub>=2.1.0
groq>=0.4.0